app/requirements.txt
# aiofiles==23.2.1
# aiohttp==3.9.1
# aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
arel==0.2.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
frozenlist==1.4.0
h11==0.14.0
# httpcore==1.0.2
# httptools==0.6.1
# httpx==0.25.2
idna==3.6
install==1.3.5
Jinja2==3.1.2
jinja2-fragments==1.2.1
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.2
pydantic-settings
pydantic_core
pytailwindcss==0.2.0
python-dotenv==1.0.0
python-multipart==0.0.6
# PyYAML==6.0.1
# replicate==0.21.1
sniffio==1.3.0
sse-starlette==1.8.2
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
watchgod==0.6
websockets==12.0
yarl==1.9.3